refactor(requirements): host dep-graph closure for shared use
`computePublishClosure` moves from `scripts/ci/connect-publish-graph.ts`
to `packages/requirements/src/dep-graph/`, re-exported from
`@trezor/requirements`. The same BFS shape already lived next door in
`requireConnectPublicDependencies.createSnapshot`; centralizing now
lets the existing `pack-packages.ts` / `connect-bump-versions.ts`
migrations (#27448) and a future custom eslint rule for missing
runtime deps consume it without copy-paste.

The walker stays pure (no fs); the FS resolver is a factory
(`createReadWorkspaceDeps(repoRoot)`) so the package itself doesn't
guess its own location — important under tsx's CJS require hook where
`import.meta.dirname` is undefined.
