+++ packages/requirements/src/dep-graph/computePublishClosure.ts
@@ -3,13 +3,15 @@
 // Filesystem-agnostic — pass a `getDeps` resolver that returns each package's
 // `dependencies` + `peerDependencies` (NOT `devDependencies`, that is the
 // publish boundary). The default filesystem-backed resolver lives in
-// `./read-workspace-deps`. Splitting them keeps this module purely
-// in-memory so it can be unit-tested without filesystem fixtures.
+// `./readWorkspaceDeps`. Splitting them keeps this module purely in-memory
+// so it can be unit-tested without filesystem fixtures.
 //
 // Single source of truth for the publish dep graph. Consumers:
 // - scripts/ci/gen-workflow-paths.ts (workflow path filters)
 // - scripts/ci/get-connect-dependencies-to-release.ts (release detection)
-// Future: pack-packages.ts and a custom eslint rule for missing-runtime-deps.
+// Future: pack-packages.ts, a custom eslint rule for missing-runtime-deps,
+// and consolidation with `requireConnectPublicDependencies` (which currently
+// implements the same BFS shape against `dependencies + optionalDependencies`).
 
 export type PackageDepsResolver = (packageName: string) => string[];
