+++ packages/requirements/src/dep-graph/index.ts
@@ -1,0 +1,3 @@
+export { computePublishClosure } from './computePublishClosure';
+export type { PackageDepsResolver } from './computePublishClosure';
+export { createReadWorkspaceDeps } from './readWorkspaceDeps';
