+++ packages/requirements/src/index.ts
@@ -3,3 +3,5 @@ export type {
     ForbiddenDependency,
     ForbiddenDepsConfig,
 } from './requirements/forbidden-deps/forbiddenDepsTypes';
+export { computePublishClosure, createReadWorkspaceDeps } from './dep-graph';
+export type { PackageDepsResolver } from './dep-graph';
