--- scripts/jest.config.cjs
@@ -1,5 +1,0 @@
-const baseConfig = require('../jest.config.base.swc');
-
-module.exports = {
-    ...baseConfig,
-};
