+++ scripts/package.json
@@ -7,13 +7,13 @@
         "depcheck": "yarn g:depcheck",
         "type-check": "yarn g:tsc --build",
         "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'",
-        "test:unit": "yarn g:jest",
         "generate-package": "tsx generatePackage.ts"
     },
     "devDependencies": {
         "@babel/cli": "7.28.6",
         "@mobily/ts-belt": "^3.13.1",
         "@trezor/eslint": "workspace:*",
+        "@trezor/requirements": "workspace:*",
         "@types/semver": "^7.7.0",
         "chalk": "^5.6.2",
         "cross-fetch": "^4.1.0",
